tpm: Fix the type of the return value in calc_tpm2_event_size()
mainline inclusion
from mainline-5.1-rc5
commit b9d0a85d6b2e
category: bugfix
bugzilla: 13838
CVE: NA

-------------------------------------------------

calc_tpm2_event_size() has an invalid signature because
it returns a 'size_t' where as its signature says that
it returns 'int'.

Cc: <stable@vger.kernel.org>
Fixes: 4d23cc323cdb ("tpm: add securityfs support for TPM 2.0 firmware event log")
Suggested-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Conflicts:
  drivers/char/tpm/eventlog/tpm2.c
[yyl: adjust context]

+++ drivers/char/tpm/eventlog/tpm2.c
@@ -37,7 +37,7 @@
  *
  * Returns size of the event. If it is an invalid event, returns 0.
  */
-static int calc_tpm2_event_size(struct tcg_pcr_event2 *event,
+static size_t calc_tpm2_event_size(struct tcg_pcr_event2 *event,
 				struct tcg_pcr_event *event_header)
 {
 	struct tcg_efi_specid_event *efispecid;
